LayoutTests/http/tests/xmlhttprequest/set-dangerous-headers-expected.txt

CONSOLE MESSAGE: line 1: Refused to set unsafe header "ACCEPT-CHARSET"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "ACCEPT-ENCODING"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "ACCESS-CONTROL-REQUEST-HEADERS"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "ACCESS-CONTROL-REQUEST-METHOD"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "AUTHORIZATION"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "CONNECTION"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "CONTENT-LENGTH"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "CONTENT-TRANSFER-ENCODING"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "COOKIE"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "COOKIE2"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "DATE"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "EXPECT"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "HOST"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "KEEP-ALIVE"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "ORIGIN"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "REFERER"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "TE"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "TRAILER"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "TRANSFER-ENCODING"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "UPGRADE"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "USER-AGENT"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "VIA"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "Proxy-"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "Proxy-test"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "PROXY-FOO"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "Sec-"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "Sec-test"
CONSOLE MESSAGE: line 1: Refused to set unsafe header "SEC-FOO"
Test that setRequestHeader cannot be used to alter security-sensitive headers.

SUCCESS
